Relocate sourcing to new file
